+++ src/update/effects.c
@@ -365,6 +365,10 @@ void run_command(UpdateResult result)
 
         case CMD_SORT_LIBRARY:
                 sort_library();
+                if (model->state.ui.current_library_sort == SORT_BY_NAME)
+                        set_error_message("Sort Alphabetically");
+                else
+                        set_error_message("Sort By Modified Date");
                 break;
 
         case CMD_VIEW_CHANGED: